fix the description of input and reformat manual hyper parameter function for local outlier factor
